271: Progress Notifications should support more than 2G of bytes sent/received
In progress notifications, bytes received and bytes sent should be Int64 and not Int32 as they contain the cumulative values of the number of bytes received or sent and not just the number immediately received or sent. As it is fully possible to upload/download more than 2G of data with Web API this should be changed to Int64 to match how we deal with content-length in general (always as Int64).

+++ src/System.Net.Http.Formatting/Handlers/HttpProgressEventArgs.cs
@@ -16,7 +16,7 @@ namespace System.Net.Http.Handlers
         /// <param name="userToken">Any user state provided as part of reading or writing the data.</param>
         /// <param name="bytesTransferred">The current number of bytes either received or sent.</param>
         /// <param name="totalBytes">The total number of bytes expected to be received or sent.</param>
-        public HttpProgressEventArgs(int progressPercentage, object userToken, int bytesTransferred, long? totalBytes)
+        public HttpProgressEventArgs(int progressPercentage, object userToken, long bytesTransferred, long? totalBytes)
             : base(progressPercentage, userToken)
         {
             BytesTransferred = bytesTransferred;
@@ -26,7 +26,7 @@ namespace System.Net.Http.Handlers
         /// <summary>
         /// Gets the current number of bytes transferred.
         /// </summary>
-        public int BytesTransferred { get; private set; }
+        public long BytesTransferred { get; private set; }
 
         /// <summary>
         /// Gets the total number of expected bytes to be sent or received. If the number is not known then this is null.
